// src/generators/chain/Types.ts
export type Gen<V> = Generator<V> | AsyncGenerator<V> | IterableIterator<V>;

/**
 * Some kind of (async) generator or an array of data of type V
 */
export type GenOrData<V> = Array<V> | Gen<V>;

/**
 * A function which can form part of a chain.
 * It takes an input {@link GenOrData}, and returns a new generator.
 */
export type Link<In, Out> = (input: GenOrData<In>) => AsyncGenerator<Out>;

/**
 * A function which can start a chain, since it takes no input
 */
export type GenFactoryNoInput<Out> = () => AsyncGenerator<Out>;
// ...
export type LinksWithSource<In, Out> = [
  Link<In, any> | GenOrData<In> | GenFactoryNoInput<In>,
  ...Array<Link<any, any>>,
  Link<any, Out>
]

/**
 * An array of chain links without a source
 */
export type Links<In, Out> = [
  Link<In, any>,
  ...Array<Link<any, any>>,
  Link<any, Out>
]


/**
 * Delay options
 */
export type DelayOptions = {
  /**
   * Time before yielding
   */
  before?: Interval,
  /**
   * Time after yielding
   */
  after?: Interval
}

export type TickOptions = {
  interval: Interval
  loops?: number
  elapsed?: Interval
  asClockTime?: boolean
}


/**
 * A rank function that compares A and B.
 * Returns the highest value, 'a' or 'b'. 
 * Returns 'eq' if values are equal
 */
export type RankFunction<T> = (a: T, b: T) => `a` | `b` | `eq`

export type RankOptions = {
  /**
   * If set, only values with this JS type are included
   */
  includeType?: `string` | `number` | `object` | `boolean`
  /**
   * If _true_, also emits values when they rank equal with current highest.
   * _false_ by default
   */
  emitEqualRanked?: boolean
  /**
   * If _true_, emits the current highest value even if it hasn't changed.
   * This means it will match the tempo of the incoming stream.
   */
  emitRepeatHighest?: boolean
}

export type RankArrayOptions = RankOptions & {
  /**
   * If _true_, it's only the highest _within_ an array that is considered,
   * rather than tracking the higest between arrays
   * Default: _false_
   */
  withinArrays?: boolean
}


/**
 * Lazy execution of a chain
 */
export type LazyChain<In, Out> = {
  /**
   * Sets `data` to be the data for the chain
   * @param data 
   * @returns
   */
  input: (data: GenOrData<In>) => LazyChain<In, Out>
  /**
   * Return the results of the chain as a regular generator.
   * If `data` is not supplied, the last value given calling `input(data)` is used.
   * @param data 
   * @returns 
   */
  asGenerator: (data?: GenOrData<In>) => AsyncGenerator<Out>
  /**
   * Returns the results of the chain as an array.
   * If `data` is not supplied, the last value given calling `input(data)` is used.
   * @param data 
   * @returns 
   */
  asArray: (data?: GenOrData<In>) => Promise<Array<Out>>
  asAsync: (data?: GenOrData<In>) => LazyChain<In, Out>
  /**
   * Gets the last output value from the chain.
   * If `data` is not supplied, the last value given calling `input(data)` is used.
   * @param data 
   * @returns 
   */
  lastOutput: (data?: GenOrData<In>) => Promise<Out | undefined>
  /**
   * Gets the first output value from the chain.
   * If `data` is not supplied, the last value given calling `input(data)` is used.
   * @param data 
   * @returns 
   */
  firstOutput: (data?: GenOrData<In>) => Promise<Out | undefined>
  /**
   * Uses a function as a source of values
   * @param callback 
   * @returns 
   */
  fromFunction: (callback: () => any) => LazyChain<any, any>
  /**
   * Take `limit` number of values from the chain before ending
   * @param limit 
   * @returns 
   */
  take: (limit: number) => LazyChain<In, Out>
  /**
   * Only emit values that have ranked higher than previously seen
   */
  rank: (r: RankFunction<In>, options: Partial<RankOptions>) => LazyChain<In, Out>
  rankArray: (r: RankFunction<In>, options: Partial<RankArrayOptions>) => LazyChain<In, Out>
  /**
   * Debounce values
   * @param duration 
   * @returns 
   */
  debounce: (duration: Interval) => LazyChain<In, Out>
  /**
   * Delay emitting values
   * @param options 
   * @returns 
   */
  delay: (options: DelayOptions) => LazyChain<In, Out>
  /**
   * Chunk values into arrays
   * @param size 
   * @param returnRemainers 
   * @returns 
   */
  chunk: (size: number, returnRemainers?: boolean) => LazyChain<In, Out>
  /**
   * Only allow values that meet `predicate` to pass
   * @param predicate 
   * @returns 
   */
  filter: (predicate: (input: any) => boolean) => LazyChain<In, Out>
  /**
   * Gets the minimum numerical value (if relevant)
   * @returns 
   */
  min: () => LazyChain<any, number>
  /**
   * Gets the maximum numerical value (if relevant)
   * @returns 
   */
  max: () => LazyChain<any, number>
  /**
   * Gets the average numerical value (if relevant)
   * @returns
   */
  average: () => LazyChain<any, number>
  /**
   * Gets the total of numerical values
   * @returns 
   */
  total: () => LazyChain<In, number>
  /**
   * Emits a running tally of how many values have been emitted
   * @returns 
   */
  tally: () => LazyChain<In, number>
  /**
   * Ignore values that match `predicate` (opposite of `filter()`)
   * @param predicate 
   * @returns 
   */
  drop: (predicate: (value: In) => boolean) => LazyChain<In, Out>
  /**
   * Emit values until `period` has elapsed
   * @param period 
   * @returns 
   */
  duration: (period: Interval) => LazyChain<In, Out>
  /**
   * Flatten values in an array into a single value
   * @param flattener 
   * @returns 
   */
  flatten: (flattener: (values: Array<any>) => any) => LazyChain<In, Out>
  /**
   * Transform an input value to an output
   * @param transformer 
   * @returns 
   */
  transform: (transformer: (v: any) => any) => LazyChain<In, Out>
}